Changed type of array to uint8_t

+++ src/LineairRegression.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include "stdint.h"
+
 class LineairRegression {
     public:
-        static float get_slope(int* yval, int count, bool* error = nullptr);
+        static float get_slope(uint8_t* yval, int count, bool* error = nullptr);
 };
